Yard greening
A program which calculate the full price for greening the full yard even with discount. Here is the steps i used to write this program: 1st - I made a variable which read a written number by user/s. 2nd - I made a variable which calculate the full price for all meters for yard greening. 3rd - I made a variable which calculate the discount from the full price. 4th - I made a variable which calculate the full price with discount. 5th - I printed the sums in sentence as asked in my assignment.