Working Hours. ex
This is a program which prints the text "open" if the written data: (hour, day of week) are versus of the graphic of the office. And the program prints the text "closed" if the written data does not equals with the hour from the graphic of the office for a working day or the written data equals with the day "Sunday".