Ski trip. ex
This program calculates the whole price which the clients of a hotel will have to pay after their sitting there. Here are some factors that define the whole sum: type of the room, nights, and what is the opinion of the client about the hotel when he going away. Here are many discounts which define the final price: